+++ client/src/filter/spatial/MapFilterContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
 import SpatialFilter from './MapFilter'
-import { toggleExcludeGlobal } from '../actions/SearchParamActions'
+import { toggleExcludeGlobal } from '../../actions/SearchParamActions'
 
 const mapStateToProps = state => {
   return {}
